@@ -6,32 +6,36 @@
   "license": "ISC",
   "main": "server.js",
   "dependencies": {
-    "assert": "^1.4.1",
+    "@material-ui/core": "^3.9.3",
+    "assert": "^1.5.0",
     "autoprefixer": "^6.5.3",
     "babel-core": "^6.18.0",
     "babel-loader": "^6.2.7",
     "babel-preset-es2015": "^6.18.0",
     "babel-preset-react": "^6.16.0",
-    "body-parser": "^1.4.3",
+    "body-parser": "^1.19.0",
+    "create-react-class": "^15.6.3",
     "css-loader": "^0.25.0",
-    "express": "^4.4.5",
+    "express": "^4.17.0",
     "extract-text-webpack-plugin": "^1.0.1",
     "flexbox": "0.0.3",
     "html-webpack-plugin": "^2.24.0",
-    "jquery": "^3.1.1",
+    "jquery": "^3.4.1",
     "jquery-ui": "^1.12.1",
-    "mongodb": "^2.2.11",
+    "mongodb": "^2.2.36",
+    "node-sass": "^4.12.0",
     "postcss-loader": "^1.0.0",
-    "react": "^15.3.2",
-    "react-dom": "^15.3.2",
+    "react": "^16.8.6",
+    "react-dom": "^16.8.6",
     "react-listbox": "^1.2.5",
     "react-router": "^3.0.0",
+    "reactcss": "^1.2.3",
     "remarkable": "^1.7.1",
     "style-loader": "^0.13.1",
     "webpack": "^1.13.3"
   },
   "devDependencies": {
-    "nodemon": "^1.11.0",
+    "nodemon": "^1.19.0",
     "webpack-dev-server": "^1.16.2"
   },
   "scripts": {
