@@ -1,31 +1,40 @@
 import React from 'react';
 import $ from 'jquery';
 import TopNav from './topnav';
+
 import Card from '@material-ui/core/Card';
 import CardContent from '@material-ui/core/CardContent';
-var createClass = require('create-react-class');
+import { Typography } from '@material-ui/core';
 import { Router, Route, browserHistory } from 'react-router';
 import { API_URL, POLL_INTERVAL } from './global';
-import { Typography } from '@material-ui/core';
+import Tabs from '@material-ui/core/Tabs';
+import Tab from '@material-ui/core/Tab';
+var createClass = require('create-react-class');
 
 module.exports = createClass({
     getInitialState: function () {
-        return { analysis: [], _isMounted: false, selectedRace: null, races: null, search: "", header: "Recently Added" };
+        return { analysis: [], _isMounted: false, currentSeason: "", selectedRace: null, races: null, search: "", header: "Archived Race Analysis", seasons: [] };
     },
     componentDidMount: function () {
         this.state._isMounted = true;
         this.loadAnalysisFromServer();
+        this.setState(
+            {
+                seasons: [
+                    "Outdoor 2019"
+                ],
+                currentSeason: "Outdoor 2019"
+            }
+        );
     },
     componentWillUnmount: function () {
         this.state._isMounted = false;
     },
     handleSearchChange: function (e) {
         this.setState({ search: e.target.value });
-        if (this.state.search == "") {
-            this.setState({ header: "Recently Added" })
-        } else {
-            this.setState({ header: "Search Results" })
-        }
+    },
+    handleSearchSeasonChange: function (e) {
+        this.setState({ searchSeason: e.target.value });
     },
     loadAnalysisFromServer: function () {
         if (this.state._isMounted) {
@@ -45,14 +54,11 @@ module.exports = createClass({
         }
     },
     expand: function (id) {
-        console.log(id);
         if (this.state.selectedRace != id) {
             this.state.selectedRace = id;
         } else {
             this.state.selectedRace = null;
         }
-
-        console.log(this.state)
         this.setState({ races: this.createHTML() });
 
     },
@@ -60,8 +66,6 @@ module.exports = createClass({
         //AJAX Request to Database searching for a meet like race and a runner name
         // Save the returned data
         //.then (
-        console.log("State Data from PastRaces:");
-        console.log(analysis);
         if (analysis != undefined && analysis != null) {
             browserHistory.push({
                 pathname: "/race",
@@ -77,6 +81,11 @@ module.exports = createClass({
         //var searchTerms = this.state.search.split(" "); // This could be useful for later
         return analysis.name.toLowerCase().includes(this.state.search.toLowerCase()) || analysis.meet.toLowerCase().includes(this.state.search.toLowerCase());
     },
+    chooseSeason: function (season) {
+        this.setState({ currentSeason: season });
+        console.log("Currently Chosen Season is " + season);
+        //AJAX Request for only archives of that season
+    },
     createHTML: function () {
         var relevantResults = this.state.analysis.filter(analysis => {
             if (this.state.search == "") {
@@ -87,50 +96,68 @@ module.exports = createClass({
         });
         // Sorts results by date and places the latest results at the top
         relevantResults.sort((a, b) => parseFloat(b.date) - parseFloat(a.date));
-        relevantResults.forEach(result => {
-            console.log(result.ID + "  " + result.date)
-        });
         return relevantResults.map(analysis => {
+            var key = analysis.name + analysis.meet + Math.random(1000);
+            var clickFuntion = () => this.navigate(analysis);
             var words = analysis.thoughts.substr(0, 40).split(" ");
             words.pop();
             var thoughts = words.join(" ") + "...";
-            return (
-                <a key={analysis.name + analysis.meet + Math.random(1000)} onClick={() => this.navigate(analysis)}>
-                    <Card className="racecard" raised= { true } >
-                        <CardContent>
-                            <Typography variant = "headline" color="primary">
-                                {analysis.name}
-                            </Typography>
-                            <Typography variant = "title">
-                                {analysis.meet} {analysis.event}
-                            </Typography>
-                            <Typography>
-                                { thoughts }
-                            </Typography>
+            return (<a key={key} onClick={clickFuntion}>
+                <Card className="racecard" raised={true} >
+                    <CardContent>
+                        <Typography variant="headline" color="primary">
+                            {analysis.name}
+                        </Typography>
+                        <Typography variant="title">
+                            {analysis.meet} {analysis.event}
+                        </Typography>
+                        <Typography>
+                            {thoughts}
+                        </Typography>
 
-                        </CardContent>
-                    </Card>
-                </a>
-            );
-
+                    </CardContent>
+                </Card> </a>);
         });
     },
     render: function () {
         this.state.races = this.createHTML();
+        var Radios = this.state.seasons.map(season => {
+            return (
+                <Tab id={season} onClick={() => this.chooseSeason(season)} name={season} label={season} value={season} />
+            );
+        });
+        const { value } = this.state;
         return (
             <div>
-                <h1>Current Season Race Analysis</h1>
+                <h1>Race Analysis Archives </h1>
                 <TopNav></TopNav>
                 <div className="container">
                     <div id="searchbar">
                         <input
                             id="name"
                             type="text"
-                            placeholder="Search..."
+                            placeholder="Search All Seasons..."
                             onChange={this.handleSearchChange}
                         />
                     </div>
-                    <h3> {this.state.header} </h3>
+                    <Tabs
+                        value={this.state.currentSeason}
+                        onChange={this.chooseSeason}
+                        indicatorColor="primary"
+                        textColor="primary"
+                        variant="scrollable"
+                        scrollButtons="auto"
+                    >
+                        {Radios}
+                    </Tabs>
+                    <div id="searchbar">
+                        <input
+                            id="name"
+                            type="text"
+                            placeholder={ "Search " + this.state.currentSeason + "..." }
+                            onChange={this.handleSearchSeasonChange}
+                        />
+                    </div>
                     <div id="racelist">
                         {this.state.races}
                     </div>
