@@ -1,9 +1,10 @@
 import React from 'react';
 import $ from 'jquery';
 import TopNav from './topnav';
+var createClass = require('create-react-class');
 import { API_URL, API_BUGS_URL, POLL_INTERVAL } from './global';
 
-module.exports = React.createClass({
+module.exports = createClass({
     getInitialState: function() {
         return {data: [], _isMounted: false, bugdesc: "", name: "", submitted:""};
     },
