@@ -1,8 +1,9 @@
 import React from 'react';
 import $ from 'jquery';
 import '../css/base.css'
+var createClass = require('create-react-class');
 import { Router, Route, browserHistory } from 'react-router';
-module.exports = React.createClass({
+module.exports = createClass({
 	
 	getInitialState: function () {
         return {runner: ''};
