@@ -1,14 +1,15 @@
 import React from 'react';
 import TopNav from './topnav';
+var createClass = require('create-react-class');
 
-module.exports = React.createClass({
-    getInitialState: function() {
-        return {data: [], _isMounted: false};
+module.exports = createClass({
+    getInitialState: function () {
+        return { data: [], _isMounted: false };
     },
-    componentDidMount: function() {
+    componentDidMount: function () {
         this.state._isMounted = true;
     },
-    componentWillUnmount: function() {
+    componentWillUnmount: function () {
         // Reset the isMounted flag so that the loadCommentsFromServer callback
         // stops requesting state updates when the commentList has been unmounted.
         // This switch is optional, but it gets rid of the warning triggered by
@@ -16,7 +17,7 @@ module.exports = React.createClass({
         // See https://reactjs.org/blog/2015/12/16/ismounted-antipattern.html
         this.state._isMounted = false;
     },
-    render: function() {
+    render: function () {
         return (
             <div>
                 <h1> Welcome to Calvin Post Race Analysis! </h1>
@@ -24,7 +25,10 @@ module.exports = React.createClass({
                 <p> Click the Analysis tab to fill out your post race form. </p>
                 <p> Click the Race Archives to see recent analysis by your teammates. </p>
                 <p> Click the Bugs tab to report a bug or suggest an improvement </p>
+                <h2> Development News</h2>
+                <p> Seperation of Current vs Archived Analysis </p>
+                
             </div>
-            );
+        );
     }
 });
