@@ -29,29 +29,32 @@ var database = 'calvinpostrace'
 var postraceDB = null;
 var mclient = require('mongodb').MongoClient
 
+// Maintenance Flags
 var sync = false;
 var addID = false;
 var clearNoID = false;
-var databaseToClearNoIDs = "backup-races"
 var addDate = false;
 var clearNoDate = false;
 var recoverFromBackup = false;
 var clearDuplicate = false;
+var seasonComplete = false;
+
 //Connect to Mongo Database
 mclient.connect(`mongodb://${username}:${password}@${host}:${port}/${database}`, function (err, client) {
     if (err) {
         console.warn(err)
         console.warn("Continuing Deployment without Database Access...")
-    } else {    
+    } else {
         postraceDB = client.db(database);
         console.log("Connected Successfully to MongoDB.")
-        if(sync) synchronizeBackup();
-        if(addID) writeNewIDs();
-        if(clearNoID) clearNoIDs(databaseToClearNoIDs);
-        if(addDate) addDates();
-        if(clearNoDate) clearNoDates("races");
-        if(recoverFromBackup) addFromBackup();
-        if(clearDuplicate) clearDuplicates("races");
+        if (sync) synchronizeBackup();
+        if(seasonComplete) archiveRaces();
+        if (addID) writeNewIDs();
+        if (clearNoID) clearNoIDs("backup-races");
+        if (addDate) addDates();
+        if (clearNoDate) clearNoDates("races");
+        if (recoverFromBackup) addFromBackup();
+        if (clearDuplicate) clearDuplicates("races");
     }
 });
 
@@ -60,63 +63,63 @@ var db;
 var APP_PATH = path.join(__dirname, 'dist');
 app.set('port', (process.env.PORT || 3000));
 app.use(bodyParser.json());
-app.use(bodyParser.urlencoded({extended: true}));
+app.use(bodyParser.urlencoded({ extended: true }));
 
 
 //Create Endpoints
 
 // GET - /api/usernames/
-app.get('/api/races', function(req, res) {
+app.get('/api/races', function (req, res) {
     var usernamesList = postraceDB.collection('races').find({}).toArray((err, result) => {
-        if(err) throw err;
+        if (err) throw err;
         res.json(result);
     });
 });
 
 // POST - /api/races/
-app.post('/api/races', function(req, res, next) {
+app.post('/api/races', function (req, res, next) {
     postraceDB.collection('races').find({}).toArray((err, array) => {
         var duplicates = array.filter(item => item.ID == req.body.ID);
-        if(duplicates.length > 0) {
+        if (duplicates.length > 0) {
 
             postraceDB.collection('races').update(duplicates[0], req.body);
             res.statusCode = 200;
-            res.send({result: "Successful Update", body: req.body, dup: duplicates});    
+            res.send({ result: "Successful Update", body: req.body, dup: duplicates });
         } else {
             postraceDB.collection('races').insert(req.body);
             res.statusCode = 200;
-            res.send({result: "Successful Insert", body: req.body, dup: duplicates});    
-        } 
+            res.send({ result: "Successful Insert", body: req.body, dup: duplicates });
+        }
     });
 });
 
 // DELETE - /api/races
-app.delete('/api/races', function(req, res, next) {
+app.delete('/api/races', function (req, res, next) {
     try {
         postraceDB.collection('races').delete(req.body);
-    } catch(e) {
+    } catch (e) {
         console.warn(e);
     }
 });
 
 // POST - /api/bugs
-app.post('/api/bugs', function(req, res, next) {
+app.post('/api/bugs', function (req, res, next) {
     postraceDB.collection('bugs').find({}).toArray((err, array) => {
         var duplicates = array.filter(item => item.name == req.body.name && item.bugdesc == req.body.bugdesc);
-        if(duplicates.length > 0) {
+        if (duplicates.length > 0) {
             postraceDB.collection('bugs').update(req.body);
             res.statusCode = 200;
-            res.send({result: "Successful Update", body: req.body, dup: duplicates});    
+            res.send({ result: "Successful Update", body: req.body, dup: duplicates });
         } else {
             postraceDB.collection('bugs').insert(req.body);
             res.statusCode = 200;
-            res.send({result: "Successful Insert", body: req.body, dup: duplicates});    
+            res.send({ result: "Successful Insert", body: req.body, dup: duplicates });
         }
     });
 });
 
 // Add Headers to responses
-app.use(function(req, res, next) {
+app.use(function (req, res, next) {
     res.setHeader('Access-Control-Allow-Origin', '*');
     res.setHeader('Cache-Control', 'no-cache');
     next();
@@ -127,20 +130,26 @@ app.use('/', express.static(APP_PATH));
 app.use('*', express.static(APP_PATH));
 
 // Launch Server
-app.listen(app.get('port'), function() {
-        console.log('Server started: http://localhost:' + app.get('port') + '/');
+app.listen(app.get('port'), function () {
+    console.log('Server started: http://localhost:' + app.get('port') + '/');
 });
+
+/////////////////////////////////
+/////////////////////////////////
+// Server Maintence Operations //
+/////////////////////////////////
+/////////////////////////////////
 
 function synchronizeBackup() {
     postraceDB.collection("races").find({}).toArray((err, array) => {
-        if(err) {
+        if (err) {
             console.warn("Backup Failed...")
             console.warn(err);
         } else {
             array.forEach(doc => {
                 var backupFile = doc;
                 delete backupFile._id;
-                postraceDB.collection("backup-races").update({ID: doc.ID}, backupFile, {upsert: true});
+                postraceDB.collection("backup-races").update({ ID: doc.ID }, backupFile, { upsert: true });
             });
             console.log("Backup completed successfully.")
         }
@@ -149,14 +158,14 @@ function synchronizeBackup() {
 
 function addFromBackup() {
     postraceDB.collection("backup-races").find({}).toArray((err, array) => {
-        if(err) {
+        if (err) {
             console.warn("Recovery Failed...")
             console.warn(err);
         } else {
             array.forEach(doc => {
                 var backupFile = doc;
                 delete backupFile._id;
-                postraceDB.collection("races").update({ID: doc.ID}, backupFile, {upsert: true});
+                postraceDB.collection("races").update({ ID: doc.ID }, backupFile, { upsert: true });
             });
             console.log("Recovery completed successfully.")
         }
@@ -165,19 +174,19 @@ function addFromBackup() {
 
 function writeNewIDs() {
     postraceDB.collection("races").find({}).toArray((err, array) => {
-        if(err) {
+        if (err) {
             console.warn("writing new IDs Failed...")
             console.warn(err);
         } else {
             array.forEach(doc => {
                 console.log("Updating Doc " + doc.name + " " + doc.meet)
-                if(doc.ID == undefined) {
+                if (doc.ID == undefined) {
                     var addID = doc;
                     delete addID._id;
                     addID.ID = doc.name + doc.meet
                     console.log("New Doc");
                     console.log(addID.ID + "\n" + addID.name + "\n" + addID.meet)
-                    postraceDB.collection("races").update(doc, addID, {upsert: true})
+                    postraceDB.collection("races").update(doc, addID, { upsert: true })
                 }
             })
             console.log("Writing New IDs Successful.")
@@ -187,17 +196,17 @@ function writeNewIDs() {
 
 function addDates() {
     postraceDB.collection("races").find({}).toArray((err, array) => {
-        if(err) {
+        if (err) {
             console.warn("Add Date failed")
             console.warn(err);
         } else {
             array.forEach(doc => {
-                if(doc.date == undefined) {
+                if (doc.date == undefined) {
                     var newDoc = doc;
                     delete newDoc._id;
                     newDoc.date = Date.now();
                     console.log("Adding Date to " + doc.name + " " + doc.meet)
-                    postraceDB.collection("races").update({ID: doc.ID}, newDoc);
+                    postraceDB.collection("races").update({ ID: doc.ID }, newDoc);
                 }
             });
             console.log("Date Adding Sucessful.");
@@ -207,14 +216,14 @@ function addDates() {
 
 function clearNoIDs(database) {
     postraceDB.collection(database).find({}).toArray((err, array) => {
-        if(err) {
+        if (err) {
             console.warn("Clear Failed...")
             console.warn(err);
         } else {
             array.forEach(doc => {
                 console.log("New Doc");
                 console.log(doc.ID + "\n" + doc.name + "\n" + doc.meet)
-                if(doc.ID == undefined) {
+                if (doc.ID == undefined) {
                     console.log("Deleting " + doc.name + " " + doc.meet)
                     postraceDB.collection(database).remove(doc)
                 } else {
@@ -228,14 +237,14 @@ function clearNoIDs(database) {
 
 function clearNoDates(database) {
     postraceDB.collection(database).find({}).toArray((err, array) => {
-        if(err) {
+        if (err) {
             console.warn("Clear Failed...")
             console.warn(err);
         } else {
             array.forEach(doc => {
                 console.log("New Doc");
                 console.log(doc.date + "\n" + doc.name + "\n" + doc.meet)
-                if(doc.date == undefined) {
+                if (doc.date == undefined) {
                     console.log("Deleting " + doc.name + " " + doc.meet)
                     postraceDB.collection(database).remove(doc)
                 } else {
@@ -249,7 +258,7 @@ function clearNoDates(database) {
 
 function clearDuplicates(database) {
     postraceDB.collection(database).find({}).toArray((err, array) => {
-        if(err) {
+        if (err) {
             console.warn("Duplicate Clear Failed...")
             console.warn(err);
         } else {
@@ -257,7 +266,7 @@ function clearDuplicates(database) {
             array.forEach(doc => {
                 console.log("New Doc");
                 console.log(doc.date + "\n" + doc.name + "\n" + doc.meet)
-                if(duplicates.includes(doc.ID)) {
+                if (duplicates.includes(doc.ID)) {
                     console.log("Deleting Copy of" + doc.name + " " + doc.meet)
                     postraceDB.collection(database).remove(doc)
                 } else {
@@ -268,4 +277,22 @@ function clearDuplicates(database) {
             console.log("Duplicate Clear Successful.")
         }
     });
+}
+
+function archiveRaces() {
+    synchronizeBackup();
+    postraceDB.collection("races").find({}).toArray((err, array) => {
+        if (err) {
+            console.warn("Archive Failed...")
+            console.warn(err);
+        } else {
+            array.forEach(doc => {
+                var backupFile = doc;
+                delete backupFile._id;
+                postraceDB.collection("archives").update({ ID: doc.ID }, backupFile, { upsert: true });
+            });
+            console.log("Archive completed successfully.")
+        }
+    });
+    postraceDB.collection("races").deleteMany({});
 }
